Put FRLocalizedString macro in its own file.
Removed the FRLocalizedString macro from FeedbackReporter.pch and placed it in a new file named FRLocalizedString.h. Added the new file to the Xcode project.  #included the file where needed.  This makes it easier to embed the feedbackreporter files into your own project instead of building feedbackreporter as a framework.