Better fix for dark mode bug
Basically reverts 65788e40.

Instead, just removed the placeholder text in the xibs in the various NSTextViews. For some reasons their presense resulted in black text on black background in dark mode.  Worse, since the workaround was to use @" " instead of @"" it resulted in different data being POSTed to the server and would actually crash a PHP script I use.